Port from SGWT 4.0p branch: ID Management Improvements - Document intended behavior, registration duration, SC<=>SGWT interactions - Call destroy() for SGWT widgets created from live SC objects as well as for   SGWT widgets created from scratch in Java - Set a property to trigger the SC Framework to do the callback into SGWT,   rather than relying on wrapping of SC Class.destroy() - Improve how SC objects are deleted when their IDs collide with SGWT widgets. - Change RefDataClass descendants to never register with ID management,   but merely validate ID availability if configured for global namespace - Add logic to correctly release auto-generated IDs in the event that SGWT   objects are destroyed before creation of the underlying SC object [gm]

+++ branches/3.1/main/src/com/smartgwt/client/core/BaseClass.java
@@ -135,31 +135,42 @@ public abstract class BaseClass {
             doInit();
             return jsObj;
         } else {
-            return getJsObj();            
+            return getJsObj();
         }
     }
 
     public static BaseClass getRef(JavaScriptObject jsObj) {
         return jsObj == null ? null : (BaseClass) JSOHelper.getAttributeAsObject(jsObj, SC.REF);
     }
-	
+
 	/**
 	 * Destroy this object.
 	 */
     public native void destroy() /*-{
-		var self = this.@com.smartgwt.client.core.BaseClass::getJsObj()();
-		if (self != null && self.__destroy) self.__destroy();
-		var id = this.@com.smartgwt.client.core.BaseClass::getID()();
-		if (id != null) {
+        var self = this.@com.smartgwt.client.core.BaseClass::getJsObj()();
+        if (self != null && self.__sgwtDestroy) {
+            delete self.__sgwtDestroy;
+            if (self.destroy) self.destroy();
+        }
+        var id = this.@com.smartgwt.client.core.BaseClass::getID()();
+        if (id != null) {
             this.@com.smartgwt.client.core.BaseClass::clearID()();
         }
+        this.@com.smartgwt.client.core.BaseClass::clearConfigRef()();
     }-*/;
 
     private void clearID() {
+        if (JSOHelper.getAttributeAsBoolean(config, "_autoAssignedID") &&
+            getRef(this.config) == null) SC.releaseID(getClass().getName(), this.id);
         IDManager.unregisterID(this, this.id);
         this.id = null;
-    	JSOHelper.setNullAttribute(config, "ID");
-    } 
+        JSOHelper.setNullAttribute(config, "ID");
+        JSOHelper.setNullAttribute(config, "_autoAssignedID");
+    }
+
+    private void clearConfigRef() {
+        JSOHelper.setNullAttribute(this.config, SC.REF);
+    }
 
     protected void error(String attribute, String value, boolean allowPostCreate) throws IllegalStateException {
         if (allowPostCreate) {
@@ -184,24 +195,30 @@ public abstract class BaseClass {
 
     protected abstract JavaScriptObject create();
 
+    private native void wrapDestroy() /*-{
+        var self = this.@com.smartgwt.client.core.BaseClass::getOrCreateJsObj()();
+        if (self && self.__sgwtDestroy == null) self.__sgwtDestroy = function () {
+            var jObj = this.__ref;
+            if (jObj != null) jObj.@com.smartgwt.client.core.BaseClass::destroy()();
+        }
+    }-*/;
+
     protected final native void doInit()/*-{
-        var self = this.@com.smartgwt.client.core.BaseClass::getJsObj()();
-        if (self) {
-            self.__destroy = self.destroy;
-            self.destroy = function() {
-                var jObj = this.__ref;
-                jObj.@com.smartgwt.client.core.BaseClass::destroy()();
-            };
-        };
+        this.@com.smartgwt.client.core.BaseClass::wrapDestroy()();
         this.@com.smartgwt.client.core.BaseClass::onInit()();
     }-*/;
 
     protected void onInit() {}
 
+    // install callbacks for a live SC object
+    protected void onBind() {
+        wrapDestroy();
+    }
+
     public String getAttribute(String attribute) {
         return getAttributeAsString(attribute);
     }
-    
+
     public native String getAttributeAsString(String property)/*-{
         var ret;
         if(this.@com.smartgwt.client.core.BaseClass::isCreated()()) {
